backend/requirements.txt
mongoengine
flask
requests
bs4
python-slugify
flask-cors